[mempool] create shared_mempool folder in crate
