+++ mempool/src/lib.rs
@@ -63,14 +63,16 @@ extern crate prometheus;
 #[cfg(feature = "fuzzing")]
 pub mod mocks;
 pub use shared_mempool::{
-    bootstrap, CommitNotification, CommitResponse, CommittedTransaction, ConsensusRequest,
-    ConsensusResponse, MempoolClientSender, SubmissionStatus, TransactionExclusion,
-    MEMPOOL_SUBSCRIBED_CONFIGS,
+    bootstrap, network,
+    types::{
+        CommitNotification, CommitResponse, CommittedTransaction, ConsensusRequest,
+        ConsensusResponse, MempoolClientSender, SubmissionStatus, TransactionExclusion,
+        MEMPOOL_SUBSCRIBED_CONFIGS,
+    },
 };
 
 mod core_mempool;
 mod counters;
-pub mod network;
 mod shared_mempool;
 
 // module op counters
